Rename crmapper -> azureconfig
