+++ service/controller/resource/azureconfig/apiserver.go
@@ -1,4 +1,4 @@
-package crmapper
+package azureconfig
 
 import (
 	"strings"
