+++ service/controller/resource/azureconfig/availability_zones.go
@@ -1,4 +1,4 @@
-package crmapper
+package azureconfig
 
 import (
 	"sort"
