+++ service/controller/resource/azureconfig/availability_zones_test.go
@@ -1,4 +1,4 @@
-package crmapper
+package azureconfig
 
 import (
 	"strconv"
