+++ service/controller/resource/azureconfig/delete.go
@@ -1,4 +1,4 @@
-package crmapper
+package azureconfig
 
 import (
 	"context"
