+++ service/controller/resource/azureconfig/error.go
@@ -1,4 +1,4 @@
-package crmapper
+package azureconfig
 
 import (
 	"github.com/giantswarm/microerror"
